06112023 _ finished 'clients' module, need to add userID to remaining processors, then test API calls passing jwt

+++ server/modules/clients/processor.js
@@ -6,7 +6,7 @@ const axios = require('axios');
 
 module.exports = ({ db }) => {
   async function getAll(options) {
-    const { limit, clientIDs, nameFirst, nameLast, phone, city, state, zip } = options;
+    const { userID, limit, clientIDs, nameFirst, nameLast, phone, city, state, zip } = options;
     const personIDs = [];
     //if filtering by clientIDs, need to get the personID for each clientID
     if (clientIDs) {
@@ -14,7 +14,8 @@ module.exports = ({ db }) => {
         const { data, error } = await db
           .from('clients')
           .select('personID')
-          .eq('clientID', clientID);
+          .filter('userID', 'eq', userID)
+          .filter('clientID', 'eq', clientID)
 
         if (error) {
           global.logger.info(`Error getting personID from clients table ${clientID}: ${error.message}`);
@@ -60,7 +61,7 @@ module.exports = ({ db }) => {
   }
 
   async function getClientByID(options) {
-    const { data, error } = await db.from('clients').select().eq('clientID', options.clientID).single();
+    const { data, error } = await db.from('clients').select().filter('userID', 'eq', options.userID).filter('clientID', 'eq', options.clientID).single();
     if (error) {
       global.logger.info(`Error getting client by ID: ${options.clientID}:${error.message}`);
       return { error: error.message };
@@ -78,6 +79,7 @@ module.exports = ({ db }) => {
       const { data, error } = await db
         .from('clients')
         .insert({
+          userID: options.userID,
           personID: person.personID,
           createdTime: new Date().toISOString(),
         })
@@ -99,13 +101,14 @@ module.exports = ({ db }) => {
   }
 
   async function update(options) {
-    const { clientID, nameFirst, nameLast, email, phone, address1, address2, city, state, zip } = options;
+    const { userID, clientID, nameFirst, nameLast, email, phone, address1, address2, city, state, zip } = options;
 
     //get the personID for the client
     const { data, error } = await db
       .from('clients')
       .select('personID')
-      .eq('clientID', clientID)
+      .filter('userID', 'eq', userID)
+      .filter('clientID', 'eq', clientID)
       .single();
 
     if (error) {
@@ -115,6 +118,7 @@ module.exports = ({ db }) => {
       //call the PATCH /persons/{personID} endpoint
       const personID = data.personID;
       const { data: personUpdateResult } = await axios.patch(`${process.env.NODE_HOST}:${process.env.PORT}/persons/${personID}`, {
+        userID,
         nameFirst,
         nameLast,
         email,
@@ -138,7 +142,7 @@ module.exports = ({ db }) => {
   }
 
   async function deleteClient(options) {
-    const { data, error } = await db.from('clients').delete().eq({ clientID: options.clientID });
+    const { data, error } = await db.from('clients').delete().filter('userID', 'eq', options.userID).filter('clientID', 'eq', options.clientID );
 
     if (error) {
       global.logger.info(`Error deleting client ${options.clientID}: ${error.message}`);
@@ -149,7 +153,7 @@ module.exports = ({ db }) => {
   }
 
   async function existsByClientID(options) {
-    const { data, error } = await db.from('clients').select('clientID').eq({ clientID: options.clientID });
+    const { data, error } = await db.from('clients').select('clientID').filter('userID', 'eq', options.userID).filter('clientID', 'eq', options.clientID);
 
     if (error) {
       global.logger.info(`Error checking if client ${options.clientID} exists: ${error.message}`);
